Refactor shortcut manager menu
